+++ Swift - The Game/Assets/Scripts/Functions/HealthBar.cs
@@ -9,7 +9,7 @@ public class HealthBar : MonoBehaviour
     public Gradient gradient;
     public Image fill;
 
-    public void SetMaxHealth(float health)
+    public void SetMaxHealth(int health)
     {
         slider.maxValue = health;
         slider.value = health;
@@ -17,7 +17,7 @@ public class HealthBar : MonoBehaviour
         //fill.color = gradient.Evaluate(1f);
     }
 
-    public void SetHealth(float health)
+    public void SetHealth(int health)
     {
         slider.value = health;
